# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Erebus Portal API",
    "version": "1.0.63",
    "description": "API documentation for the Vampire: The Masquerade RPG portal."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "API Server"
    }
  ],
  "paths": {
...
    "/api/characters/xp/spend": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "example": "any"
                  },
                  "target": {
                    "example": "any"
                  },
                  "currentLevel": {
                    "example": "any"
                  },
                  "newLevel": {
                    "example": "any"
                  },
                  "ritualLevel": {
                    "example": "any"
                  },
                  "formulaLevel": {
                    "example": "any"
                  },
                  "dots": {
                    "example": "any"
                  },
                  "disciplineKind": {
                    "example": "any"
                  },
                  "patchSheet": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/comms/status": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/admin/comms/status": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "comms_enabled": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/comms/config": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/admin/comms/schedule": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "schedule": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}